requirements.txt
python-dotenv
pandas
numpy
psycopg2
SQLAlchemy
